+++ src/main/java/insane96mcp/iguanatweaksreborn/module/experience/enchantments/EnchantmentsFeature.java
@@ -105,7 +105,7 @@ public class EnchantmentsFeature extends JsonFeature {
 	public static Boolean preventFarmlandTramplingWithFeatherFalling = true;
 	@Config
 	@Label(name = "Replace protection enchantments", description = "If true, vanilla protection enchantments are replaced with mod's ones. To re-enable vanilla enchantments refer to `disabled_enchantments.json`.\n" +
-			"Protection has only one level, protects 6% per level and is treasure. Other protections work the same except for projectile that reduces the sight range of mobs by 2% per level. Feather falling protects for 16% per level instead of 12%.")
+			"Protection has only one level, protects 6% per level and is treasure. Other protections work the same except for projectile that reduces the sight range of mobs by 2% per level. Feather falling protects for 16% per level instead of 12% + 1 per level.")
 	public static Boolean replaceProtectionEnchantments = true;
 	@Config
 	@Label(name = "Replace damaging enchantments", description = """