+++ changelog.md
@@ -4,6 +4,7 @@
 * Increased chance for chickens, cows and sheep to drop food with Looting (Luck)
 * Reduced Pigs food drop with looting
 * Reduced Golden Tools durability and increased efficiency
+* Protection enchantment now also protects from Magic damage
 * Fixed mobs with bows dealing almost no damage
 
 ## 3.10.2